Move UI components to other files
