+++ src/styles/components/UI-game/UI-game.scss
@@ -43,6 +43,10 @@
   backdrop-filter: blur(18px);
 }
 
+.mess {
+  z-index: 2;
+}
+
 .message-container {
   position: relative;
   width: 95%;
@@ -258,7 +262,7 @@
     width: 50%;
     textarea {
       width: calc(100% - 29%);
-      height: 50px;
+      height: 110px;
     }
   }
   .message-scroll {
@@ -289,7 +293,7 @@
     width: 50%;
     textarea {
       width: calc(100% - 29%);
-      height: 50px;
+      height: 110px;
     }
   }
   .message-scroll {
