+++ src/styles/components/buttons.scss
@@ -329,7 +329,7 @@ $button-login-shadow: rgba(0, 0, 0, 0.25);
 
   #send {
     right: 140px;
-    bottom: 80px;
+    bottom: 65px;
     transform: scale(1.5);
   }
 
@@ -412,7 +412,7 @@ $button-login-shadow: rgba(0, 0, 0, 0.25);
   }
 
   #send {
-    bottom: 75px;
+    bottom: 60px;
     right: 120px;
     transform: scale(1.2);
   }
